feat: add InvalidIdException for delete in CompanyController with test case

+++ src/main/java/com/rest/springbootemployee/controller/CompanyController.java
@@ -74,7 +74,10 @@ public class CompanyController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteCompany(@PathVariable String id) {
+    public void deleteCompany(@NonNull @PathVariable String id) {
+        if(!ObjectId.isValid(id)){
+            throw new InvalidIdException();
+        }
         companyService.delete(id);
     }
 }
